requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6

# DSPy and LLM dependencies
dspy-ai==2.4.9
openai==1.3.7

# Embeddings
sentence-transformers==2.2.2
torch==2.1.1

# Database
duckdb==0.9.2

# Utilities
numpy==1.26.2
pandas==2.1.3
pyyaml==6.0.1
python-Levenshtein==0.23.0
python-dotenv==1.0.0

# Logging and observability
structlog==23.2.0
